+++ app/api/v1/schemas/grade.go
@@ -1,13 +1,13 @@
 package schemas
 
 type GradType struct {
-	Id          int64   `json:"id"`
+	Id          int     `json:"id"`
 	Description string  `json:"description"`
 	Grade       float64 `json:"grade"`
 }
 
 type GradeRes struct {
-	Id            int64      `json:"id"`
+	Id            int        `json:"id"`
 	Date          string     `json:"date"`
 	GradesPerType []GradType `json:"gradesPerType"`
 }
