+++ app/api/v1/schemas/user.go
@@ -1,8 +1,8 @@
 package schemas
 
 type User struct {
-	Id             int64  `json:"id"`
-	ClassStartYear int64  `json:"classStartYear"`
+	Id             int    `json:"id"`
+	ClassStartYear int    `json:"classStartYear"`
 	Name           string `json:"name"`
 	Email          string `json:"email"`
 	Role           string `json:"role"`
